apply review

+++ db/migrate/20240628171109_archive_users_by_organisation_rather_than_by_department.rb
@@ -14,7 +14,9 @@ class ArchiveUsersByOrganisationRatherThanByDepartment < ActiveRecord::Migration
   end
 
   def migrate_current_archives_to_organisations
-    Archive.find_each do |archive|
+    # we do that to avoid the archive created in the loop to be processed by the find_each
+    start_time = Time.current
+    Archive.where('created_at < ?', start_time).find_each do |archive|
       user = archive.user
       department = Department.find(archive.department_id)
       organisations_where_user_should_be_archived = department.organisations & user.organisations
@@ -27,11 +29,14 @@ class ArchiveUsersByOrganisationRatherThanByDepartment < ActiveRecord::Migration
   end
 
   def migrate_current_archives_to_departments
-    Archive.find_each do |archive|
+    start_time = Time.current
+    Archive.where('created_at < ?', start_time).find_each do |archive|
       organisation = Organisation.find(archive.organisation_id)
-      Archive.create(user: archive.user, department: organisation.department,
-                     archiving_reason: archive.archiving_reason, created_at: archive.created_at)
-      archive.destroy!
+      if Archive.find_by(user_id: archive.user_id, department_id: organisation.department_id).present?
+        archive.destroy!
+      else
+        archive.update!(department_id: organisation.department_id)
+      end
     end
   end
 end
